+++ skins/weewx-wdc/src/js/diagrams/d3/components/legend.ts
@@ -1,9 +1,11 @@
 import * as d3 from "d3";
-import { Serie } from "../../types";
+import { Serie, legendPosition } from "../../types";
 
 export const addLegend = (
   svgElement: d3.Selection<SVGGElement, unknown, null, undefined>,
   width: number,
+  height: number,
+  legendPosition: legendPosition,
   data: Serie[],
   units: string[],
   colors: string[],
@@ -15,11 +17,12 @@ export const addLegend = (
 
   const size = 14,
     x = 144,
-    y = 1.5;
+    y = 1.5,
+    positionsLeft = ["top left", "bottom left"];
 
   const legend = svgElement.append("g").attr("class", "legend");
 
-  data.map((item, index) => {
+  data.map((item, index): void => {
     const legendItem = legend
       .append("g")
       .attr("transform", `translate(0, ${index * 21})`);
@@ -37,7 +40,7 @@ export const addLegend = (
       .data([item.id])
       .enter()
       .append("rect")
-      .attr("x", x)
+      .attr("x", positionsLeft.includes(legendPosition) ? 0 : x)
       .attr("y", y)
       .attr("width", size)
       .attr("data-testid", `legend-rect-${item.observation}`)
@@ -56,7 +59,7 @@ export const addLegend = (
       )
       .enter()
       .append("text")
-      .attr("x", x - 10)
+      .attr("x", positionsLeft.includes(legendPosition) ? size + 5 : x - 10)
       .attr("y", y + 6)
       .style("fill", () => {
         return colors[index];
@@ -64,20 +67,64 @@ export const addLegend = (
       .text(function (d) {
         return d;
       })
-      .attr("text-anchor", "end")
+      .attr(
+        "text-anchor",
+        positionsLeft.includes(legendPosition) ? "start" : "end"
+      )
       .style("dominant-baseline", "central")
       .style("pointer-events", "none")
       .style("font-size", "11px");
   });
 
-  legend.attr(
-    "transform",
-    `translate(${
-      width -
-      (legend.node()?.getBBox().width as number) -
-      (legend.node()?.getBBox().x as number)
-    }, 0)`
-  );
+  // Legend position.
+  switch (legendPosition) {
+    case "top right":
+      legend.attr(
+        "transform",
+        `translate(${
+          width -
+          (legend.node()?.getBBox().width as number) -
+          (legend.node()?.getBBox().x as number)
+        }, 0)`
+      );
+      break;
+    case "top left":
+      legend.attr("transform", `translate(0, 0)`);
+      break;
+    case "bottom right":
+      legend.attr(
+        "transform",
+        `translate(${
+          width -
+          (legend.node()?.getBBox().width as number) -
+          (legend.node()?.getBBox().x as number)
+        }, ${
+          height -
+          (legend.node()?.getBBox().height as number) -
+          (legend.node()?.getBBox().y as number)
+        })`
+      );
+      break;
+    case "bottom left":
+      legend.attr(
+        "transform",
+        `translate(0, ${
+          height -
+          (legend.node()?.getBBox().height as number) -
+          (legend.node()?.getBBox().y as number)
+        })`
+      );
+      break;
+    default:
+      legend.attr(
+        "transform",
+        `translate(${
+          width -
+          (legend.node()?.getBBox().width as number) -
+          (legend.node()?.getBBox().x as number)
+        }, 0)`
+      );
+  }
 
   return legend;
 };